feat(view): removed dev code
