+++ src/batch.js
@@ -44,30 +44,6 @@ function batchFn(fn) {
   return batched;
 }
 
-/* CJW
-function batchMethodCallbacks(obj, method) {
-  const descriptor = Object.getOwnPropertyDescriptor(obj, method);
-  if (
-    descriptor &&
-    descriptor.writable &&
-    typeof descriptor.value === 'function'
-  ) {
-    obj[method] = new Proxy(descriptor.value, {
-      apply(target, ctx, args) {
-        return Reflect.apply(target, ctx, args.map(batchFn));
-      },
-    });
-  }
-}
-*/
-
-/* CJW
-// batched obj.addEventListener(cb) like callbacks
-function batchMethodsCallbacks(obj, methods) {
-  methods.forEach(method => batchMethodCallbacks(obj, method));
-}
-*/
-
 function batchMethod(obj, method) {
   const descriptor = Object.getOwnPropertyDescriptor(obj, method);
   if (!descriptor) {
@@ -91,50 +67,3 @@ export function batchMethods(obj, methods) {
   methods.forEach(method => batchMethod(obj, method));
   return obj;
 }
-
-// do a sync batching for the most common task sources
-// this should be removed when React's own batching is improved in the future
-
-/* CJW
-// batch timer functions
-batchMethodsCallbacks(globalObj, [
-  'setTimeout',
-  'setInterval',
-  'requestAnimationFrame',
-  'requestIdleCallback',
-]);
-*/
-
-/* CJW
-if (globalObj.Promise) {
-  batchMethodsCallbacks(Promise.prototype, ['then', 'catch']);
-}
-*/
-
-// Event listener batching causes an input caret jumping bug:
-// https://github.com/RisingStack/react-easy-state/issues/92.
-// This part has to be commented out to prevent that bug.
-// React batches setStates in its event listeners anyways
-// so this commenting this part out is not a huge issue.
-
-// batch addEventListener calls
-/* if (globalObj.EventTarget) {
-  batchMethodsCallbacks(EventTarget.prototype, [
-    'addEventListener',
-    'removeEventListener',
-  ]);
-} */
-
-/* CJW
-// this batches websocket event handlers
-if (globalObj.WebSocket) {
-  batchMethods(WebSocket.prototype, [
-    'onopen',
-    'onmessage',
-    'onerror',
-    'onclose',
-  ]);
-}
-*/
-
-// HTTP event handlers are usually wrapped by Promises, which is covered above
